Add callback to projects/ endpoint in discover component
An initial call back to the backend is added in the discover component
in order to populate cards with project data. Axios is added to handle
the callback, and some basic configuration infrastructure is added to
handle backend config for localhost/production. Abstraction of the
callback from the component is not attempted at this time, deferring to
if/when it is needed