Add 'follow' function to project cards
A followed-projects provider is added to the project to provide a single
point of context for the application to access a user's followed
projects. Each card refers to this context to avoid a callback to the
backend for each card displayed. At this time, access to this context is
not abstracted from the cards so they can be more generally used; this can be
done at such time as that is needed. A blank login page is added to
provide a landing point for follow clicks when a user is anonymous.