Add a login page
The login page skeleton is filled in using the existing hero section
layout (which is abstracted to be reused) and the existing login form.
Additional props are added to optionally control the cancel behaviour
and redirectOnLogin behaviour. This page does not currently redirect
when you are already logged in, which will be added when control over
page accessibility is added more widely