Move logged-in-user state to a provider
The logged-in-user state is useful across many pages beyond the
navigation, where it was initially managed. To make this more widely
available, it is moved to a provider, and a global provider component is
created to encapsulate this provider and the followed-projects provider.
This also means that we can have a single point where the user state is
managed and thus manage the storage of the user data in local state &
its subsequent removal on logout - meaning we can have a consistant
appearance in the navbar when the user is logged in; instead of a
flicker in the avatar while we wait for the user the be queried. Logout
logic is encapsulated in the provider and should always be logged out
via this method henceforth